Move user setup stuff into one file
